:recycle: :white_check_mark: Cleanup tests - Remove unnecessary sequential tests - Move all tests with parallel gateways to the test file for parallel gateways - Remove redundant process models
