--- _integration_tests/bpmn/manual_task_sequential_test.bpmn
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<bpmn:definitions xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" id="Definition_1" targetNamespace="http://bpmn.io/schema/bpmn" exporter="BPMN Studio" exporterVersion="1">
-  <bpmn:collaboration id="Collaboration_1cidyxu" name="">
-    <bpmn:participant id="Participant_0px403d" name="manual_task_sequential_test" processRef="manual_task_sequential_test" />
-  </bpmn:collaboration>
-  <bpmn:process id="manual_task_sequential_test" name="manual_task_sequential_test" isExecutable="true">
-    <bpmn:laneSet>
-      <bpmn:lane id="Lane_1xzf0d3" name="Default_Test_Lane">
-        <bpmn:flowNodeRef>StartEvent_1</bpmn:flowNodeRef>
-        <bpmn:flowNodeRef>EndEvent_0eie6q6</bpmn:flowNodeRef>
-        <bpmn:flowNodeRef>Manual_Task_1</bpmn:flowNodeRef>
-        <bpmn:flowNodeRef>Manual_Task_2</bpmn:flowNodeRef>
-      </bpmn:lane>
-    </bpmn:laneSet>
-    <bpmn:startEvent id="StartEvent_1" name="Start Event">
-      <bpmn:outgoing>SequenceFlow_0vlhj88</bpmn:outgoing>
-    </bpmn:startEvent>
-    <bpmn:endEvent id="EndEvent_0eie6q6" name="End Event">
-      <bpmn:incoming>SequenceFlow_0pwmcoa</bpmn:incoming>
-    </bpmn:endEvent>
-    <bpmn:sequenceFlow id="SequenceFlow_0vlhj88" sourceRef="StartEvent_1" targetRef="Manual_Task_1" />
-    <bpmn:manualTask id="Manual_Task_1" name="ManualTask 1">
-      <bpmn:incoming>SequenceFlow_0vlhj88</bpmn:incoming>
-      <bpmn:outgoing>SequenceFlow_0u9jra6</bpmn:outgoing>
-    </bpmn:manualTask>
-    <bpmn:sequenceFlow id="SequenceFlow_0u9jra6" sourceRef="Manual_Task_1" targetRef="Manual_Task_2" />
-    <bpmn:manualTask id="Manual_Task_2" name="ManualTask 2">
-      <bpmn:incoming>SequenceFlow_0u9jra6</bpmn:incoming>
-      <bpmn:outgoing>SequenceFlow_0pwmcoa</bpmn:outgoing>
-    </bpmn:manualTask>
-    <bpmn:sequenceFlow id="SequenceFlow_0pwmcoa" sourceRef="Manual_Task_2" targetRef="EndEvent_0eie6q6" />
-  </bpmn:process>
-  <bpmndi:BPMNDiagram id="BPMNDiagram_1">
-    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="Collaboration_1cidyxu">
-      <bpmndi:BPMNShape id="Participant_0px403d_di" bpmnElement="Participant_0px403d">
-        <dc:Bounds x="5" y="4" width="581" height="170" />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="Lane_1xzf0d3_di" bpmnElement="Lane_1xzf0d3">
-        <dc:Bounds x="35" y="4" width="551" height="170" />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="StartEvent_1mox3jl_di" bpmnElement="StartEvent_1">
-        <dc:Bounds x="83" y="69" width="36" height="36" />
-        <bpmndi:BPMNLabel>
-          <dc:Bounds x="74" y="105" width="55" height="14" />
-        </bpmndi:BPMNLabel>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="EndEvent_0eie6q6_di" bpmnElement="EndEvent_0eie6q6">
-        <dc:Bounds x="503" y="69" width="36" height="36" />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNEdge id="SequenceFlow_0vlhj88_di" bpmnElement="SequenceFlow_0vlhj88">
-        <di:waypoint x="119" y="87" />
-        <di:waypoint x="169" y="87" />
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNShape id="ManualTask_0vezkdx_di" bpmnElement="Manual_Task_1">
-        <dc:Bounds x="169" y="47" width="100" height="80" />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNEdge id="SequenceFlow_0u9jra6_di" bpmnElement="SequenceFlow_0u9jra6">
-        <di:waypoint x="269" y="87" />
-        <di:waypoint x="319" y="87" />
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNShape id="ManualTask_0lopaty_di" bpmnElement="Manual_Task_2">
-        <dc:Bounds x="319" y="47" width="100" height="80" />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNEdge id="SequenceFlow_0pwmcoa_di" bpmnElement="SequenceFlow_0pwmcoa">
-        <di:waypoint x="419" y="87" />
-        <di:waypoint x="503" y="87" />
-      </bpmndi:BPMNEdge>
-    </bpmndi:BPMNPlane>
-  </bpmndi:BPMNDiagram>
-</bpmn:definitions>
