--- _integration_tests/bpmn/user_task_sequential_test.bpmn
@@ -1,92 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<bpmn:definitions xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:camunda="http://camunda.org/schema/1.0/bpmn" id="Definitions_12t7o2o" targetNamespace="http://bpmn.io/schema/bpmn" exporter="Camunda Modeler" exporterVersion="1.15.1">
-  <bpmn:collaboration id="Collaboration_05evzp4">
-    <bpmn:participant id="Participant_005yfyb" name="user_task_sequential_test" processRef="user_task_sequential_test" />
-  </bpmn:collaboration>
-  <bpmn:process id="user_task_sequential_test" name="user_task_sequential_test" isExecutable="true">
-    <bpmn:laneSet>
-      <bpmn:lane id="Lane_12gp2ih" name="Default_Test_Lane">
-        <bpmn:flowNodeRef>StartEvent_1</bpmn:flowNodeRef>
-        <bpmn:flowNodeRef>EndEvent_15v0olt</bpmn:flowNodeRef>
-        <bpmn:flowNodeRef>User_Task_1</bpmn:flowNodeRef>
-        <bpmn:flowNodeRef>User_Task_2</bpmn:flowNodeRef>
-      </bpmn:lane>
-    </bpmn:laneSet>
-    <bpmn:startEvent id="StartEvent_1" name="StartEvent_1&#10;">
-      <bpmn:outgoing>SequenceFlow_1akei4o</bpmn:outgoing>
-    </bpmn:startEvent>
-    <bpmn:endEvent id="EndEvent_15v0olt" name="End">
-      <bpmn:incoming>SequenceFlow_068tsl3</bpmn:incoming>
-    </bpmn:endEvent>
-    <bpmn:sequenceFlow id="SequenceFlow_1akei4o" sourceRef="StartEvent_1" targetRef="User_Task_1" />
-    <bpmn:sequenceFlow id="SequenceFlow_068tsl3" sourceRef="User_Task_2" targetRef="EndEvent_15v0olt" />
-    <bpmn:sequenceFlow id="SequenceFlow_13n8mus" sourceRef="User_Task_1" targetRef="User_Task_2" />
-    <bpmn:userTask id="User_Task_1" name="User Task 1">
-      <bpmn:extensionElements>
-        <camunda:formData>
-          <camunda:formField id="Sample_Form_Field" type="string" />
-        </camunda:formData>
-      </bpmn:extensionElements>
-      <bpmn:incoming>SequenceFlow_1akei4o</bpmn:incoming>
-      <bpmn:outgoing>SequenceFlow_13n8mus</bpmn:outgoing>
-    </bpmn:userTask>
-    <bpmn:userTask id="User_Task_2" name="User Task 2">
-      <bpmn:extensionElements>
-        <camunda:formData>
-          <camunda:formField id="Sample_Form_Field" />
-        </camunda:formData>
-      </bpmn:extensionElements>
-      <bpmn:incoming>SequenceFlow_13n8mus</bpmn:incoming>
-      <bpmn:outgoing>SequenceFlow_068tsl3</bpmn:outgoing>
-    </bpmn:userTask>
-  </bpmn:process>
-  <bpmndi:BPMNDiagram id="BPMNDiagram_1">
-    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="Collaboration_05evzp4">
-      <bpmndi:BPMNShape id="Participant_005yfyb_di" bpmnElement="Participant_005yfyb">
-        <dc:Bounds x="78" y="158" width="778" height="301" />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="Lane_12gp2ih_di" bpmnElement="Lane_12gp2ih">
-        <dc:Bounds x="108" y="158" width="748" height="301" />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="StartEvent_08bx2ed_di" bpmnElement="StartEvent_1">
-        <dc:Bounds x="178" y="289" width="36" height="36" />
-        <bpmndi:BPMNLabel>
-          <dc:Bounds x="164" y="328" width="64" height="25" />
-        </bpmndi:BPMNLabel>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="EndEvent_15v0olt_di" bpmnElement="EndEvent_15v0olt">
-        <dc:Bounds x="741" y="289" width="36" height="36" />
-        <bpmndi:BPMNLabel>
-          <dc:Bounds x="749" y="328" width="20" height="13" />
-        </bpmndi:BPMNLabel>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNEdge id="SequenceFlow_1akei4o_di" bpmnElement="SequenceFlow_1akei4o">
-        <di:waypoint x="214" y="307" />
-        <di:waypoint x="287" y="307" />
-        <bpmndi:BPMNLabel>
-          <dc:Bounds x="250.5" y="285.5" width="0" height="13" />
-        </bpmndi:BPMNLabel>
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNEdge id="SequenceFlow_068tsl3_di" bpmnElement="SequenceFlow_068tsl3">
-        <di:waypoint x="626" y="307" />
-        <di:waypoint x="741" y="307" />
-        <bpmndi:BPMNLabel>
-          <dc:Bounds x="683.5" y="285.5" width="0" height="13" />
-        </bpmndi:BPMNLabel>
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNEdge id="SequenceFlow_13n8mus_di" bpmnElement="SequenceFlow_13n8mus">
-        <di:waypoint x="387" y="307" />
-        <di:waypoint x="526" y="307" />
-        <bpmndi:BPMNLabel>
-          <dc:Bounds x="456.5" y="285.5" width="0" height="13" />
-        </bpmndi:BPMNLabel>
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNShape id="UserTask_08zfr3m_di" bpmnElement="User_Task_1">
-        <dc:Bounds x="287" y="267" width="100" height="80" />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="UserTask_0f9wr8u_di" bpmnElement="User_Task_2">
-        <dc:Bounds x="526" y="267" width="100" height="80" />
-      </bpmndi:BPMNShape>
-    </bpmndi:BPMNPlane>
-  </bpmndi:BPMNDiagram>
-</bpmn:definitions>
