+++ _integration_tests/test/empty_activity_tests.js
@@ -49,76 +49,6 @@ describe('Empty Activity - ', () => {
     });
   });
 
-  it('should finish two sequential EmptyActivities', async () => {
-
-    const startEventIdToUse = 'StartEvent_2';
-    const correlationId = uuid.v4();
-    const samplePayload = {
-      test: 'value',
-    };
-
-    testFixtureProvider.executeProcess(processModelId, startEventIdToUse, correlationId, samplePayload);
-    await processInstanceHandler.waitForProcessInstanceToReachSuspendedTask(correlationId, processModelId);
-
-    let waitingEmptyActivities = await processInstanceHandler.getWaitingEmptyActivitiesForCorrelationId(identity, correlationId);
-
-    should(waitingEmptyActivities.emptyActivities).be.instanceOf(Array);
-    should(waitingEmptyActivities.emptyActivities.length).be.greaterThan(0);
-
-    const emptyActivity1 = waitingEmptyActivities.emptyActivities[0];
-
-    await processInstanceHandler
-      .finishEmptyActivityInCorrelation(identity, emptyActivity1.processInstanceId, correlationId, emptyActivity1.flowNodeInstanceId);
-    await processInstanceHandler.waitForProcessInstanceToReachSuspendedTask(correlationId);
-
-    waitingEmptyActivities = await processInstanceHandler.getWaitingEmptyActivitiesForCorrelationId(identity, correlationId);
-
-    should(waitingEmptyActivities.emptyActivities).be.instanceOf(Array);
-    should(waitingEmptyActivities.emptyActivities.length).be.greaterThan(0);
-
-    const emptyActivity2 = waitingEmptyActivities.emptyActivities[0];
-
-    return new Promise(async (resolve, reject) => {
-      processInstanceHandler.waitForProcessWithInstanceIdToEnd(emptyActivity2.processInstanceId, resolve);
-      await processInstanceHandler
-        .finishEmptyActivityInCorrelation(
-          identity,
-          emptyActivity2.processInstanceId,
-          emptyActivity2.correlationId,
-          emptyActivity2.flowNodeInstanceId,
-        );
-    });
-  });
-
-  it('should finish two parallel running EmptyActivities', async () => {
-
-    const startEventIdToUse = 'StartEvent_3';
-    const correlationId = uuid.v4();
-    const samplePayload = {
-      test: 'value',
-    };
-
-    testFixtureProvider.executeProcess(processModelId, startEventIdToUse, correlationId, samplePayload);
-    await processInstanceHandler.waitForProcessInstanceToReachSuspendedTask(correlationId, processModelId, 2);
-
-    const waitingEmptyActivities = await processInstanceHandler.getWaitingEmptyActivitiesForCorrelationId(identity, correlationId);
-
-    should(waitingEmptyActivities).have.property('emptyActivities');
-    should(waitingEmptyActivities.emptyActivities).have.size(2, 'There should be two waiting EmptyActivities');
-
-    const emptyActivities = waitingEmptyActivities.emptyActivities;
-
-    return new Promise(async (resolve, reject) => {
-      processInstanceHandler.waitForProcessWithInstanceIdToEnd(emptyActivities[0].processInstanceId, resolve);
-
-      for (const emptyActivity of emptyActivities) {
-        await testFixtureProvider
-          .consumerApiService
-          .finishEmptyActivity(identity, emptyActivity.processInstanceId, correlationId, emptyActivity.flowNodeInstanceId);
-      }
-    });
-  });
-
   it('should fail to finish a non existing EmptyActivity', async () => {
 
     const correlationId = uuid.v4();
